moved severity enum to constants
