+++ src/Constants.ts
@@ -77,3 +77,10 @@ export const HARDCODED_RULES = {
 		category: 'Scanner Internal'
 	}
 };
+
+export enum Severity {
+	NONE = 0,
+	LOW = 3,
+	MODERATE = 2,
+	HIGH = 1
+}
