--- src/lib/util/Severity.ts
@@ -1,7 +1,0 @@
-
-export enum Severity {
-	NONE = 0,
-	LOW = 3,
-	MODERATE = 2,
-	HIGH = 1
-}
